Crate rain

Visualize vertical data inside your terminal

This library helps you to display line based data vertically within your terminal. The color of the actual graph represents its value, whereas blue is low and red is high. These color bounds will be calculated automatically during runtime. Beside this, the terminal dimensions are adapted during runtime, too. If no data was added to a line, their terminal line is dashed.

Example usage

use rain::Graph;

// Get a drawing area
let mut graph = Graph::new();

// Get some line identifiers
let l1 = "Line 1";
let l2 = "Line 1";
let l3 = "Line 1";

// Add some values and print
assert!(graph.add(l1, 0).is_ok());
assert!(graph.add(l2, 0).is_ok());
graph.print();

// Add more values and print
assert!(graph.add(l2, 5).is_ok());
assert!(graph.add(l3, 10).is_ok());
graph.print();

// Remove a line and print
assert!(graph.remove(l1).is_ok());
graph.print();

Modules

error
Basic error handling mechanisms

Structs

Graph
The graph drawing structure